pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurogen"
version = "0.1.0"
description = "Cellular Automata Developmental Initialization for Language Models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "tiktoken>=0.5.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
meta = ["cma>=3.3.0"]
search = ["optuna>=3.4.0"]
datasets = ["datasets>=2.14.0"]
monitoring = ["tensorboard>=2.14.0"]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
all = ["neurogen[meta,datasets,monitoring,dev]"]

[tool.setuptools.packages.find]
include = ["neurogen*"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "slow: marks tests that take >30s",
    "gpu: marks tests requiring CUDA GPU",
    "mps: marks tests requiring Apple MPS backend",
    "integration: marks end-to-end integration tests",
]
